逻辑改为动态选择流量占比选择最优转化率的阈值，增加行业大图平均cvr的比较

+++ src/main/scala/com/cpc/spark/shortvideo/shortvideo.scala
@@ -446,7 +446,7 @@ group by searchid, adtype,userid,ideaid,isclick,isreport,exp_cvr_ori,
     var cnt = 0
     for (record <- tabfinal2.collect()) {
       var userid = record.getAs[String]("userid")
-      var exp_cvr = record.getAs[Int]("expcvr")
+      var exp_cvr = record.getAs[Long]("expcvr")
       println(s"""useridr:$userid, expcvr:${exp_cvr}""")
 
       cnt += 1
